+++ hefquin-access/src/main/java/se/liu/ida/hefquin/federation/access/FederationAccessManager.java
@@ -19,6 +19,26 @@ public interface FederationAccessManager extends StatsProvider
 	CompletableFuture<CardinalityResponse> issueCardinalityRequest( ReqType req, MemberType fm )
 			throws FederationAccessException;
 
+	default < ReqType extends DataRetrievalRequest,
+	          RespType extends DataRetrievalResponse<?>,
+	          MemberType extends FederationMember >
+	CompletableFuture<RespType> issueRequest( ReqType req, MemberType fm, boolean ignoreCache )
+			throws FederationAccessException
+	{
+		return issueRequest(req, fm);
+	};
+
+	default < ReqType extends DataRetrievalRequest,
+	          RespType extends DataRetrievalResponse<?>,
+	          MemberType extends FederationMember >
+	CompletableFuture<CardinalityResponse> issueCardinalityRequest( ReqType req,
+	                                                                MemberType fm,
+	                                                                boolean ignoreCardinalityCache )
+			throws FederationAccessException
+	{
+		return issueCardinalityRequest(req, fm);
+	}
+
 	@Override
 	FederationAccessStats getStats();