+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/executable/impl/ops/ExecOpLookupJoinViaWrapperWithParamVars.java
@@ -172,7 +172,7 @@ public class ExecOpLookupJoinViaWrapperWithParamVars
 
 			final CompletableFuture<StringResponse> f;
 			try {
-				f = ctx.getFederationAccessMgr().issueRequest(req, fm);
+				f = ctx.getFederationAccessMgr().issueRequest( req, fm, ctx.ignoreCache() );
 			}
 			catch ( final FederationAccessException e ) {
 				log.debug( "Request issuance failed for endpoint {} (paramValues={})", fm, entry.getKey() );